+++ core/include/string_utils.hpp
@@ -1,6 +1,5 @@
 
 
-
 #include <string>
 
 namespace Drill
@@ -8,5 +7,7 @@ namespace Drill
     namespace string_utils
     {
         bool tokenSearch(std::string main, std::string tokens);
-    }
-}
+
+        std::string to_lower(const std::string &);
+    } // namespace string_utils
+} // namespace Drill
