+++ core/src/string_utils.cpp
@@ -2,9 +2,11 @@
 
 #include <string_utils.hpp>
 
-#include <vector>
-#include <string>
+#include <algorithm>
+#include <cctype>
 #include <sstream>
+#include <string>
+#include <vector>
 
 namespace Drill
 {
@@ -28,18 +30,31 @@ namespace Drill
             return elems;
         }
 
-        bool tokenSearch(std::string main, std::string tokens_string) { 
-            auto tokens_vector = split(tokens_string, ' ');
+        bool tokenSearch(std::string main, std::string tokens_string)
+        {
+            auto tokens_vector = split(to_lower(tokens_string), ' ');
+
+            auto main_lower = to_lower(main);
 
             // if all tokens no matter the order are inside main it's okay
-            for (auto token : tokens_vector) {
-            
+            for (auto token : tokens_vector)
+            {
+
                 // if token not found
-                if (main.find(token) == std::string::npos) {
+                if (main_lower.find(token) == std::string::npos)
+                {
                     return false;
                 }
             }
             return true;
         }
+
+        std::string to_lower(const std::string &str)
+        {
+            std::string result = str;
+            std::transform(str.begin(), str.end(), result.begin(), ::tolower);
+            return result;
+        }
     } // namespace string_utils
+
 } // namespace Drill
